Userservice

+++ src/main/java/br/com/dcx/ufpb/eng/ApiCiclismo/service/UserService.java
@@ -18,22 +18,37 @@ public class UserService {
     }
 
     public User updateUser(Long userId, User user) {
-        // 
-        return userRepository.save(user);
+        Optional<User> existingUserOptional = userRepository.findById(userId);
+        if (existingUserOptional.isPresent()) {
+            User existingUser = existingUserOptional.get();
+            
+            // Atualizar os campos necessários do usuário existente com os valores do novo usuário
+            existingUser.setUsername(user.getUsername());
+            existingUser.setEmail(user.getEmail());
+            existingUser.setFullName(user.getFullName());
+            
+            return userRepository.save(existingUser);
+        } else {
+            throw new UserNotFoundException("User not found with ID: " + userId);
+        }
     }
 
     public void deleteUser(Long userId) {
-        // 
-        userRepository.deleteById(userId);
+        // Verificar se o usuário com o ID fornecido existe no banco de dados
+        Optional<User> userOptional = userRepository.findById(userId);
+        if (userOptional.isPresent()) {
+            userRepository.deleteById(userId);
+        } else {
+            throw new UserNotFoundException("User not found with ID: " + userId);
+        }
     }
 
     public User getUserById(Long userId) {
-        // 
-        return userRepository.findById(userId).orElse(null);
+         return userRepository.findById(userId)
+            .orElseThrow(() -> new UserNotFoundException("User not found with ID: " + userId));
     }
 
     public List<User> getAllUsers() {
-        // 
         return userRepository.findAll();
     }
 }
